+++ src/controller/python/ChipDeviceController-StorageDelegate.cpp
@@ -29,7 +29,7 @@
 namespace chip {
 namespace Controller {
 
-CHIP_ERROR PythonPersistentStorageDelegate::SyncGetKeyValue(const char * key, void * value, uint16_t & size)
+CHIP_ERROR PythonPersistentStorageDelegate::SyncGetKeyValue(const char * key, void * value, size_t & size)
 {
     auto val = mStorage.find(key);
     if (val == mStorage.end())
@@ -60,7 +60,7 @@ CHIP_ERROR PythonPersistentStorageDelegate::SyncGetKeyValue(const char * key, vo
     return CHIP_NO_ERROR;
 }
 
-CHIP_ERROR PythonPersistentStorageDelegate::SyncSetKeyValue(const char * key, const void * value, uint16_t size)
+CHIP_ERROR PythonPersistentStorageDelegate::SyncSetKeyValue(const char * key, const void * value, size_t size)
 {
     mStorage[key] = std::string(static_cast<const char *>(value), size);
     ChipLogDetail(Controller, "SyncSetKeyValue: %s=%s", key, value);
