+++ src/controller/python/ChipDeviceController-StorageDelegate.h
@@ -36,8 +36,8 @@ class PythonPersistentStorageDelegate : public PersistentStorageDelegate
 {
 public:
     PythonPersistentStorageDelegate() {}
-    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, uint16_t & size) override;
-    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, uint16_t size) override;
+    CHIP_ERROR SyncGetKeyValue(const char * key, void * value, size_t & size) override;
+    CHIP_ERROR SyncSetKeyValue(const char * key, const void * value, size_t size) override;
     CHIP_ERROR SyncDeleteKeyValue(const char * key) override;
 
 private:
